adding sounds and ui improvements
